requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.0.3
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.19.2
psycopg2
SQLAlchemy==1.4.36
Werkzeug==2.1.2

